ml/requirements.txt
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
httpx>=0.27.0
